Student

+++ Uniexam/src/main/java/it/unical/uniexam/hibernate/dao/StudentDAO.java
@@ -12,10 +12,11 @@ import it.unical.uniexam.hibernate.domain.utility.PhoneNumber;
 
 public interface StudentDAO {
 	
-	public Long addStundent(String name, String surname, String password, Address address, Set<Email> emails, Set<PhoneNumber> phoneNumbers, DegreeCourse degreeCourse_registered, Long serialNumber);
+	public Long addStundent(String name, String surname, String password, Address address, Set<Email> emails, Set<PhoneNumber> phoneNumbers, DegreeCourse degreeCourse_registered, int serialNumber);
 	public Set<Student> getStudents();
 	public Student getStudent(Long serialNumber);
 	
 	public Set<Course> getCarrier(Long serialNumber);
 	public Set<Group> getGroups(Long serialNumber);	
 }
+
